pyproject.toml
[project]
name = "archivist-rag"
version = "0.1.0"
description = "RAG system for personal Markdown knowledge base"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "streamlit>=1.40.0",
    "python-frontmatter>=1.1.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
ingest = "scripts.ingest:main"
